+++ R/piping-model.R
@@ -1,8 +1,10 @@
+
 #' @export
 #' @rdname model
 model.function <- function(x, ..., append=FALSE, auto=TRUE, envir=parent.frame()) {
   .modelLines <- .quoteCallInfoLines(match.call(expand.dots = TRUE)[-(1:2)], envir=envir)
   .ret <- rxUiDecompress(rxode2(x))
+  if (length(.modelLines) == 0) return(.ret$modelFun)
   .modelHandleModelLines(.modelLines, .ret, modifyIni=FALSE, append=append, auto=auto, envir=envir)
 }
 
@@ -11,7 +13,18 @@ model.function <- function(x, ..., append=FALSE, auto=TRUE, envir=parent.frame()
 model.rxUi <- function(x, ..., append=FALSE, auto=TRUE, envir=parent.frame()) {
   .modelLines <- .quoteCallInfoLines(match.call(expand.dots = TRUE)[-(1:2)], envir=envir)
   .ret <- rxUiDecompress(.copyUi(x)) # copy so (as expected) old UI isn't affected by the call
-  .modelHandleModelLines(.modelLines, .ret, modifyIni=FALSE, append=append, auto=auto, envir=envir)
+  if (length(.modelLines) == 0) return(.ret$modelFun)
+  .ret <- .modelHandleModelLines(.modelLines, .ret, modifyIni=FALSE, append=append, auto=auto, envir=envir)
+  # need to adjust since the model function was from a rxui object
+  .x <- rxUiDecompress(x)
+  .ret <- rxUiDecompress(.ret)
+  .ret <- .newModelAdjust(.ret, .x)
+  .ret <- rxUiCompress(.ret)
+  .cls <- setdiff(class(x), class(.ret))
+  if (length(.cls) > 0) {
+    class(.ret) <- c(.cls, class(.ret))
+  }
+  .ret
 }
 
 #' @export
@@ -20,6 +33,7 @@ model.rxode2 <- function(x, ..., append=FALSE, auto=TRUE, envir=parent.frame()) 
   .modelLines <- .quoteCallInfoLines(match.call(expand.dots = TRUE)[-(1:2)], envir=envir)
   x <- as.function(x)
   .ret <- rxUiDecompress(rxode2(x))
+  if (length(.modelLines) == 0) return(.ret$modelFun)
   .modelHandleModelLines(.modelLines, .ret, modifyIni=FALSE, append=append, auto=auto, envir=envir)
 }