+++ R/rxUiBlessed.R
@@ -1,0 +1,8 @@
+## created by .createRxUiBlessedList() in function-body.R edit there
+.rxUiBlessed <- c("covLhsDf", "covariates", "curHi", "curLow", 
+    "errParams0", "estNotAllowed", "eta", "etaLhsDf", "extraCmt", 
+    "extraDvid", "hasErrors", "iniDf", "lstExpr", "meta", "model", 
+    "modelName", "muRefCovariateDataFrame", "muRefCovariateEmpty", 
+    "muRefCurEval", "muRefDataFrame", "muRefDropParameters", 
+    "muRefExtra", "muRefExtraEmpty", "mv0", "mvL", "nonMuEtas", 
+    "oneTheta", "predDf", "sticky", "thetaLhsDf")